third_party/WebKit/LayoutTests/inspector/tracing/timeline-recalculate-styles-expected.txt

Tests the Timeline API instrumentation of a style recalculation event

Test data
RecalculateStyles Properties:
{
    data : {
        elementCount : 2
        frame : <string>
    }
    endTime : <number>
    stackTrace : <object>
    startTime : <number>
    thread : <string>
    type : "RecalculateStyles"
}
